gh-26 Turn gmrc into namespace package and rename backend
