+++ setup.py
@@ -4,15 +4,15 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="gmrc",  # Replace with your own username
-    version="0.0.0",
+    name="gmrc",
+    version="0.0.1-dev",
     author="CLTL",
-    # author_email="t.kim@vu.nl",
-    # description="face bounding box detection, 68 landmarks, age, gender, 512-D embeddings",
+    author_email="piek.vossen@vu.nl",
+    description="Representation of multi-modal datasets",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    # url="https://github.com/leolani/cltl-face-all",
-    packages=setuptools.find_packages(),
+    url="https://github.com/cltl/GMRCAnnotation",
+    packages=['gmrc.representation'],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
